+++ lib/src/filter/base_filter.dart
@@ -38,6 +38,7 @@ abstract class PMFilter {
   /// Construct a default filter.
   factory PMFilter.defaultValue({
     bool containsPathModified = false,
+    bool includeHiddenAssets = false,
   }) {
     return CustomFilter.sql(
       where: '',
@@ -57,6 +58,17 @@ abstract class PMFilter {
   /// See also:
   ///  * [AssetPathEntity.lastModified].
   bool containsPathModified = false;
+
+  /// Whether to include hidden assets in the results.
+  ///
+  /// This option only takes effect on iOS.
+  /// Beginning with iOS 16, users can require authentication to view the
+  /// hidden album, and the user setting is true by default. When true,
+  /// the system doesn’t return hidden assets even if the option is true.
+  ///
+  /// See also:
+  ///  * [PHFetchOptions.includeHiddenAssets](https://developer.apple.com/documentation/photos/phfetchoptions/includehiddenassets).
+  bool includeHiddenAssets = false;
 
   /// The type of the filter.
   BaseFilterType get type;
@@ -83,6 +95,7 @@ abstract class PMFilter {
   Map<String, dynamic> _paramMap() {
     return <String, dynamic>{
       'containsPathModified': containsPathModified,
+      'includeHiddenAssets': includeHiddenAssets,
     };
   }
 }