+++ CHANGELOG.md
@@ -10,9 +10,17 @@ To know more about breaking changes, see the [Migration Guide][].
 
 ### Features
 
+- Add `includeHiddenAssets` option to `PMFilter` and `FilterOptionGroup` to include hidden assets in the results on iOS.
 - Add `cancelToken` parameter to `AssetEntity.loadFile`.
 - Add `cancelAllRequest` method to `PhotoManager`.
 - The `getFile` and `getOriginBytes` methods are public.
+
+### Fixes
+
+- Fix `SizeConstraint` not working for videos on Android. (#1275)
+- Fix inconsistent resizing behavior between iOS and Android in `thumbnailDataWithSize`. (#1271)
+
+## 3.6.4
 
 ### Fixes